+++ app/src/main/java/com/techcourse/service/UserService.java
@@ -1,39 +1,12 @@
 package com.techcourse.service;
 
-import com.interface21.jdbc.datasource.DataSourceTransactionManager;
-import com.interface21.jdbc.datasource.TransactionTemplate;
-import com.techcourse.dao.UserDao;
-import com.techcourse.dao.UserHistoryDao;
 import com.techcourse.domain.User;
-import com.techcourse.domain.UserHistory;
 
-public class UserService {
+public interface UserService {
 
-    private final UserDao userDao;
-    private final UserHistoryDao userHistoryDao;
-    private final TransactionTemplate transactionTemplate;
+    User findById(final long id);
 
-    public UserService(final UserDao userDao, final UserHistoryDao userHistoryDao,
-                       final DataSourceTransactionManager transactionManager) {
-        this.userDao = userDao;
-        this.userHistoryDao = userHistoryDao;
-        this.transactionTemplate = new TransactionTemplate(transactionManager);
-    }
+    void insert(final User user);
 
-    public User findById(final long id) {
-        return userDao.findById(id);
-    }
-
-    public void insert(final User user) {
-        userDao.insert(user);
-    }
-
-    public void changePassword(final long id, final String newPassword, final String createBy) {
-        transactionTemplate.execute(() -> {
-            final var user = findById(id);
-            user.changePassword(newPassword);
-            userDao.update(user);
-            userHistoryDao.log(new UserHistory(user, createBy));
-        });
-    }
+    void changePassword(final long id, final String newPassword, final String createBy);
 }
